Fix improper nested lexical scoping of a variable
This meant that the prefix match did not actually inform the search
location for the quote.

Since I would prefer to be defensive here about making such a big
change after shipping 1.0, I think it's wise to search for the first
slice with high distance tolerance even if we found the prefix. In
this manner, we will hopefully find quotes when their prefixes are
no longer nearby with greater probability but still get the benefit
of disambiguation by prefix.