Match short strings more strictly when missing context
In a scenario where we have a prefix and a suffix but fail to find them,
we should be skeptical that the text is actually going to be in the
document. In such a scenario, short selections are problematic, because
they are likely to match against the document purely by chance.

This commit:

- Adds support for finding the suffix and adjusting the search location
  accordingly.
- Drops the diff-match-patch Match_Threshold (i.e. the laxness) to 0.25
  if the exact string is less than 32 characters long and prefix or
  suffix have failed to match.

This absolutely doesn't cover all cases -- very short strings ("a",
"the") will continue to match with high probability even if the prefix
and suffix fail -- but it does fix one common problem case.